requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
PyPDF2